requirements/base.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
base58==2.1.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cytoolz==0.11.2
ecdsa==0.15
eth-abi==2.1.1
eth-hash==0.3.2
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
httpcore==0.12.3
httpx==0.16.1
idna==3.3
multidict==6.0.2
parsimonious==0.8.1
pycryptodome==3.14.1
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
toolz==0.11.2
tronpy==0.2.6
urllib3==1.26.9
yarl==1.7.2
